Export `GDataConstructor`.
Why, because it is exposed in the type signature of `mkSumType`. So it
must be mentioned in the constraints of any polymorphic function that
does somethnig with `mkSumType`.

Example:

```Haskell
makeType :: forall thing. _ => SumType Haskell
makeType = mkSumType (Proxy @thing)
```

Without this patch, there is no way to fill in the blank!